requirements.txt
asyncio
websockets==13.1
nest_asyncio
python_dotenv
ib_insync==0.9.86
pandas
requests
requests_cache
yfinance[nospam]
pyrate_limiter
logging